fix 2 bugs
Please notice that this commit is done after i finished my work on the
assignment. Please ignore it if you want to see the result achieved in
the given amount of time.

I noticed 2 bugs during testing with 100 concurrent connections:

1. Service panics when it receives more than 10 concurrent requests
while it has no cats images in the cache

2. When rate limit for the cats API is reached, error returned together
with cached cat image is not nil.